# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a provider resolver for did ev",
        "version": "1.0.0",
        "title": "KT Provider resolver did ev",
        "contact": {
            "email": "jmirandr@emeal.nttdata.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths" : {
        "/identifiers/{identifier}" : {
            "get" : {
                "summary" : "Resolve a DID.",
                "description": "<p>This endpoint resolves a DID. As input it takes the DID, plus DID resolution <a href='https://www.w3.org/TR/did-core/#did-resolution-options'>options</a>. The output is a DID document in one of the supported representations, plus metadata.</p> <p>See the <a href='https://w3c-ccg.github.io/did-resolution/'>DID Resolution</a> specification for additional details.</p>",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "The DID to be resolved.",
                        "required": true,
                        "type": "string",
                        "examples": {
                            "example1": {
                                "value": "did:ev:cwMLAqQCguxLzd1biFQH4xpy2M7BZXvvcXKZ7",
                                "description": "A DID using the `ev` method."
                            }
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "The requested media type of the DID document representation or DID resolution result. See <a href='https://www.w3.org/TR/did-core/#representations'>https://www.w3.org/TR/did-core/#representations</a> and <a href='https://w3c-ccg.github.io/did-resolution/#did-resolution-result'>https://w3c-ccg.github.io/did-resolution/#did-resolution-result</a>.",
                        "required": false,
                        "type": "string",
                        "examples": {
                            "application/did+json": {
                                "value": "application/json",
                                "description": "Media type of a DID document (JSON representation)."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully resolved!",
                        "content": {
                            "application/did+json": {
                                "schema": {
                                    "type": "object",
                                    "description": "The DID document (JSON representation).",
                                    "example": {
                                        "@context": "https://w3id.org/did/v1",
                                        "id": "did:ev:2uzPtwJmXbBqMmP9DkR7dE3FcLmgYejdJ42",
                                        "controller": "did:ev:2uzPtwJmXbBqMmP9DkR7dE3FcLmgYejdJ42",
                                        "authentication": [{
                                            "id": "did:ev:2uzPtwJmXbBqMmP9DkR7dE3FcLmgYejdJ42#keys-1",
                                            "type": "EcdsaSecp256k1RecoveryMethod2020",
                                            "blockchainAccountId": "eip155:1:0xaeaefd50a2c5cda393e9a1eef2d6ba23f2c4fd6d"
                                        }]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid input!"
                    },
                    "500": {
                        "description": "error!"
                    }
                }
            }
        }
    }
}